Fix handling of file open mode
* Mat_VarDelete: Disable deleting variables if file is opened in read-only mode
* Mat_Create4: Fix file open mode
* Add Mat_GetFileAccessMode

As reported by https://github.com/tbeu/matio/issues/194

+++ src/mat73.c
@@ -2601,7 +2601,7 @@ Mat_Create73(const char *matname, const char *hdr_str)
     if ( err >= 116 )
         mat->header[115] = '\0'; /* Just to make sure it's NULL terminated */
     memset(mat->subsys_offset, ' ', 8);
-    mat->version = (int)0x0200;
+    mat->version = 0x0200;
     endian = 0x4d49;
 
     version = 0x0200;
